Revert "Vvartemenko/mob"
