+++ build.gradle
@@ -1,7 +1,7 @@
 plugins {
     id 'java'
     id 'io.qameta.allure' version '2.11.2'
-    id 'io.freefair.lombok' version '8.6'
+    id "io.freefair.lombok" version "6.0.0-m2"
 }
 
 allure {
@@ -16,11 +16,6 @@ allure {
             }
         }
     }
-}
-
-java {
-    sourceCompatibility = '17'
-    targetCompatibility = '17'
 }
 
 compileJava {
@@ -39,18 +34,16 @@ repositories {
 
 dependencies {
     testImplementation(
-            "org.junit.jupiter:junit-jupiter:5.10.1",
-            "com.codeborne:selenide:6.13.0",
-            "io.qameta.allure:allure-selenide:2.25.0",
-            "io.rest-assured:rest-assured:5.4.0",
-            "io.qameta.allure:allure-rest-assured:2.21.0",
-            "commons-io:commons-io:2.11.0",
-            "org.slf4j:slf4j-simple:2.0.7",
-            "org.aeonbits.owner:owner:1.0.12",
-            "org.assertj:assertj-core:3.24.2",
-            "org.apache.commons:commons-text:1.10.0",
-            "io.rest-assured:json-schema-validator:5.4.0",
-            "io.appium:java-client:8.3.0"
+            'com.codeborne:selenide:6.19.1',
+            'org.junit.jupiter:junit-jupiter:5.10.1',
+            'org.apache.commons:commons-text:1.10.0',
+            'com.github.javafaker:javafaker:1.0.2',
+            'io.qameta.allure:allure-selenide:2.25.0',
+            'io.qameta.allure:allure-rest-assured:2.21.0',
+            'org.aeonbits.owner:owner:1.0.12',
+            'io.rest-assured:rest-assured:5.4.0',
+            'org.slf4j:slf4j-simple:2.0.7',
+            'io.rest-assured:json-schema-validator:5.4.0'
     )
 }
 
@@ -67,7 +60,7 @@ tasks.withType(Test) {
 }
 
 task withTags(type: Test) {
-    String tags = System.getProperty("tags", "smokeWeb")
+    String tags = System.getProperty("tags", "smoke")
     useJUnitPlatform {
         systemProperties(System.getProperties())
         includeTags tags.split(',')
