--- src/test/java/ru/perekrestok/mob/config/AuthConfig.java
@@ -1,14 +1,0 @@
-package ru.perekrestok.mob.config;
-
-import org.aeonbits.owner.Config;
-
-@Config.Sources({
-        "classpath:mob/config/auth.properties"
-})
-public interface AuthConfig extends Config {
-    @Config.Key("user.username")
-    String getUsername();
-
-    @Config.Key("user.accessKey")
-    String getPassword();
-}
