--- src/test/java/ru/perekrestok/mob/helpers/Attach.java
@@ -1,21 +1,0 @@
-package ru.perekrestok.mob.helpers;
-
-import io.qameta.allure.Attachment;
-
-import java.nio.charset.StandardCharsets;
-
-import static com.codeborne.selenide.WebDriverRunner.getWebDriver;
-
-public class Attach {
-    @Attachment(value = "Page source", type = "text/plain")
-    public static byte[] pageSource() {
-        return getWebDriver().getPageSource().getBytes(StandardCharsets.UTF_8);
-    }
-
-    @Attachment(value = "Video", type = "text/html", fileExtension = ".html")
-    public static String addVideo(String sessionId) {
-        return "<html><body><video width='100%' height='100%' controls autoplay><source src='"
-                + Browserstack.videoUrl(sessionId)
-                + "' type='video/mp4'></video></body></html>";
-    }
-}
