--- src/test/java/ru/perekrestok/mob/helpers/Browserstack.java
@@ -1,23 +1,0 @@
-package ru.perekrestok.mob.helpers;
-
-import ru.perekrestok.mob.config.AuthConfig;
-import org.aeonbits.owner.ConfigFactory;
-
-import static io.restassured.RestAssured.given;
-
-public class Browserstack {
-    static AuthConfig authConfig = ConfigFactory.create(AuthConfig.class, System.getProperties());
-
-    public static String videoUrl(String sessionId) {
-        String url = String.format("https://api.browserstack.com/app-automate/sessions/%s.json", sessionId);
-
-        return given()
-                .auth().basic(authConfig.getUsername(), authConfig.getPassword())
-                .get(url)
-                .then()
-                .log().status()
-                .log().body()
-                .statusCode(200)
-                .extract().path("automation_session.video_url");
-    }
-}
