--- src/test/java/ru/perekrestok/mob/tests/BaseTest.java
@@ -1,40 +1,0 @@
-package ru.perekrestok.mob.tests;
-
-import com.codeborne.selenide.Configuration;
-import com.codeborne.selenide.Selenide;
-import com.codeborne.selenide.logevents.SelenideLogger;
-import ru.perekrestok.mob.drivers.BrowserstackDriver;
-import ru.perekrestok.mob.helpers.Attach;
-import io.qameta.allure.selenide.AllureSelenide;
-import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.BeforeAll;
-import org.junit.jupiter.api.BeforeEach;
-
-import static com.codeborne.selenide.Selenide.closeWebDriver;
-import static com.codeborne.selenide.Selenide.open;
-
-public class BaseTest {
-
-    @BeforeAll
-    public static void browserSetUp() {
-        System.setProperty("deviceHost", System.getProperty("deviceHost", "browserstack"));
-        Configuration.browser = BrowserstackDriver.class.getName();
-        Configuration.browserSize = null;
-        Configuration.timeout = 30000;
-    }
-
-    @BeforeEach
-    public void addLogger() {
-        SelenideLogger.addListener("allure", new AllureSelenide());
-        open();
-    }
-
-    @AfterEach
-    public void addAttachments() {
-        String sessionId = Selenide.sessionId().toString();
-        Attach.pageSource();
-        closeWebDriver();
-        if (System.getProperty("deviceHost").equals("browserstack"))
-            Attach.addVideo(sessionId);
-    }
-}
