+++ src/test/java/ru/perekrestok/web/config/DriverConfig.java
@@ -3,7 +3,7 @@ package ru.perekrestok.web.config;
 import org.aeonbits.owner.Config;
 
 @Config.Sources({
-        "classpath:web/config/${env}.properties"
+        "classpath:config/${env}.properties"
 })
 public interface DriverConfig extends Config {
     @Key("browser.name")
