+++ src/test/java/ru/perekrestok/web/tests/AddToCartTests.java
@@ -23,7 +23,7 @@ import static io.qameta.allure.Allure.step;
 @Epic("Perekrestok web")
 @Story("Корзина")
 @Feature("Добавление товара в корзину")
-@Tag("smokeWeb")
+@Tag("smoke")
 public class AddToCartTests extends BaseTest {
     private final Header header = new Header();
     private final ReceivingMethodPopup receivingMethodPopup = new ReceivingMethodPopup();
