+++ src/test/java/ru/perekrestok/web/tests/SearchTests.java
@@ -20,7 +20,7 @@ import static java.lang.String.format;
 @Epic("Perekrestok web")
 @Story("Поиск")
 @Feature("Поиск товара")
-@Tag("smokeWeb")
+@Tag("smoke")
 public class SearchTests extends BaseTest {
     private final MainPage mainPage = new MainPage();
     private final Header header = new Header();
