+++ src/test/resources/config/local.properties
@@ -1,0 +1,4 @@
+browser.name=chrome
+browser.version=117.0
+browser.size=1366x768
+is.remote=false
