--- src/test/resources/mob/apps/note.txt
@@ -1,2 +1,0 @@
-Instruction to upload app to browserstack - https://www.browserstack.com/docs/app-automate/api-reference/appium/apps#upload-an-app
-Browserstack will automatically remove the app after 30 days after uploading
