--- src/test/resources/mob/config/auth.properties
@@ -1,2 +1,0 @@
-user.username=vladislavartyome_BiuOID
-user.accessKey=5fV8e2NQctrkzoRrNL6H
