--- src/test/resources/mob/config/browserstack.properties
@@ -1,4 +1,0 @@
-browserstack.device=Google Pixel 3a
-browserstack.deviceOsVersion=9.0
-browserstack.app=Perekrestok
-browserstack.remoteUrl=https://hub.browserstack.com/wd/hub
