--- src/test/resources/web/config/local.properties
@@ -1,4 +1,0 @@
-browser.name=chrome
-browser.version=117.0
-browser.size=1366x768
-is.remote=false
