--- src/test/resources/web/config/remote.properties
@@ -1,5 +1,0 @@
-browser.name=chrome
-browser.version=122.0
-browser.size=1920x1080
-is.remote=true
-browser.remote.url=https://user1:1234@selenoid.autotests.cloud/wd/hub
